+++ Examples/GLImage Demo/GLViewExample/GLViewExampleAppDelegate.h
@@ -10,7 +10,7 @@
 
 @interface GLViewExampleAppDelegate : NSObject <UIApplicationDelegate>
 
-@property (nonatomic, retain) IBOutlet UIWindow *window;
-@property (nonatomic, retain) IBOutlet UIViewController *viewController;
+@property (nonatomic, strong) IBOutlet UIWindow *window;
+@property (nonatomic, strong) IBOutlet UIViewController *viewController;
 
 @end