+++ Examples/GLImage Demo/GLViewExample/GLViewImageExampleController.h
@@ -11,17 +11,17 @@
 
 @interface GLViewImageExampleController : UIViewController
 
-@property (nonatomic, retain) IBOutlet UIScrollView *scrollView;
-@property (nonatomic, retain) IBOutlet UIView *contentView;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView1;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView2;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView3;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView4;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView5;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView6;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView7;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView8;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView9;
-@property (nonatomic, retain) IBOutlet GLImageView *imageView10;
+@property (nonatomic, strong) IBOutlet UIScrollView *scrollView;
+@property (nonatomic, strong) IBOutlet UIView *contentView;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView1;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView2;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView3;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView4;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView5;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView6;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView7;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView8;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView9;
+@property (nonatomic, strong) IBOutlet GLImageView *imageView10;
 
 @end