+++ Examples/GLImage Demo/GLViewExample/GLViewImageMapController.m
@@ -56,10 +56,5 @@
     return cell;
 }
 
-- (void)dealloc
-{
-    [_imageMap release];
-    [super dealloc];
-}
 
 @end