+++ Examples/GLImage Demo/GLViewExample/GLViewMipmapController.h
@@ -1,0 +1,16 @@
+//
+//  GLViewMipmapController.h
+//  GLImageDemo
+//
+//  Created by Nick Lockwood on 01/12/2012.
+//  Copyright 2012 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "GLImageView.h"
+
+@interface GLViewMipmapController : UIViewController
+
+@property (nonatomic, strong) IBOutlet GLImageView *mipmapImageView;
+
+@end