+++ GLView/Core/GLView.h
@@ -2,7 +2,7 @@
 //  GLView.h
 //
 //  GLView Project
-//  Version 1.4
+//  Version 1.5 beta
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design
@@ -40,6 +40,7 @@
 @property (nonatomic, assign) CGFloat fov;
 @property (nonatomic, assign) CGFloat near;
 @property (nonatomic, assign) CGFloat far;
+@property (nonatomic, assign) CATransform3D contentTransform;
 
 - (void)setUp;
 - (void)display;