+++ GLView/Images/GLImage.h
@@ -2,7 +2,7 @@
 //  GLImage.h
 //
 //  GLView Project
-//  Version 1.4
+//  Version 1.5 beta
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design